+++ assume/strategies/learning_advanced_orders.py
@@ -10,10 +10,10 @@ import torch as th
 from assume.common.base import SupportsMinMax
 from assume.common.market_objects import MarketConfig, Orderbook, Product
 from assume.common.utils import get_products_index
-from assume.strategies.learning_strategies import RLStrategy
+from assume.strategies.learning_strategies import LearningEOMPowerplant
 
 
-class RLAdvancedOrderStrategy(RLStrategy):
+class LearningProfileEOMPowerplant(LearningEOMPowerplant):
     """
     Reinforcement Learning Strategy for an Energy-Only-Market with simple hourly, block and linked orders.