+++ assume/world.py
@@ -157,7 +157,7 @@ class World:
         self.dst_components = demand_side_technologies
 
         self.bidding_strategies = bidding_strategies
-        if "learning_powerplant" not in bidding_strategies:
+        if "learning_eom_powerplant" not in bidding_strategies:
             logger.info(
                 "Learning Strategies are not available. Check that you have torch installed."
             )