+++ tests/test_standard_powerplant_strategies.py
@@ -10,9 +10,9 @@ import pytest
 
 from assume.common.forecasts import NaiveForecast
 from assume.strategies import (
-    EOMPowerplant,
-    NCRMPowerplant,
-    PCRMPowerplant,
+    StandardEOMPowerplant,
+    StandardNCRMPowerplant,
+    StandardPCRMPowerplant,
 )
 from assume.units import PowerPlant
 
@@ -43,7 +43,7 @@ def power_plant() -> PowerPlant:
 
 def test_eom_powerplant(mock_market_config, power_plant):
     end = datetime(2023, 7, 1, 1)
-    strategy = EOMPowerplant()
+    strategy = StandardEOMPowerplant()
     mc = mock_market_config
     product_tuples = [(start, end, None)]
     bids = strategy.calculate_bids(power_plant, mc, product_tuples=product_tuples)
@@ -74,7 +74,7 @@ def test_eom_powerplant(mock_market_config, power_plant):
 
 def test_flexable_pos_reserve(mock_market_config, power_plant):
     end = datetime(2023, 7, 1, 4, 0, 0)
-    strategy = PCRMPowerplant()
+    strategy = StandardPCRMPowerplant()
     mc = mock_market_config
     mc.product_type = "energy_pos"
     product_tuples = [(start, end, None)]
@@ -99,7 +99,7 @@ def test_flexable_pos_reserve(mock_market_config, power_plant):
 
 def test_flexable_neg_reserve(mock_market_config, power_plant):
     end = datetime(2023, 7, 1, 4, 0, 0)
-    strategy = NCRMPowerplant()
+    strategy = StandardNCRMPowerplant()
     mc = mock_market_config
     # Calculations for negative energy
     mc.product_type = "energy_neg"