+++ tests/test_units_operator.py
@@ -163,8 +163,8 @@ async def test_formulate_bids(units_operator: UnitsOperator):
 
 @pytest.mark.require_learning
 async def test_write_learning_params(rl_units_operator: RLUnitsOperator):
-    from assume.strategies.learning_advanced_orders import RLAdvancedOrderStrategy
-    from assume.strategies.learning_strategies import RLStrategy
+    from assume.strategies.learning_advanced_orders import LearningProfileEOMPowerplant
+    from assume.strategies.learning_strategies import LearningEOMPowerplant
 
     marketconfig = rl_units_operator.available_markets[0]
     start = datetime(2020, 1, 1)
@@ -173,7 +173,7 @@ async def test_write_learning_params(rl_units_operator: RLUnitsOperator):
 
     params_dict = {
         "bidding_strategies": {
-            "EOM": RLAdvancedOrderStrategy(
+            "EOM": LearningProfileEOMPowerplant(
                 unit_id="testplant",
                 learning_mode=True,
             )
@@ -212,7 +212,7 @@ async def test_write_learning_params(rl_units_operator: RLUnitsOperator):
 
     rl_units_operator.units["testplant"].bidding_strategies[
         "EOM"
-    ].bidding_strategies = RLStrategy(
+    ].bidding_strategies = LearningEOMPowerplant(
         unit_id="testplant",
         learning_mode=True,
         observation_dimension=50,